Struct sndjvu_format::ser::SerializeElements

pub struct SerializeElements<'co, 'wr: 'co> { /* private fields */ }

Serializer for the elements of a DJVU or DJVI component.

Implementations

impl<'co, 'wr: 'co> SerializeElements<'co, 'wr>

pub fn anta(&mut self, ant: &str) -> Result<(), Error>

Serialize an ANTa chunk.

The DjVu standard notes that “the use of the ANTa chunk is discouraged”, the compressed ANTz chunk being preferred (see Self::antz).

See AnnotBuf for a strongly-typed way to build up data for the ant argument.

pub fn antz(&mut self, bzz: &[u8]) -> Result<(), Error>

Serialize an ANTz chunk.

See AnnotBuf for a strongly-typed way to build up data that can be compressed for the bzz argument.

pub fn txta(&mut self, text: &str, zones: &[Zone]) -> Result<(), Error>

Serialize a TXTa chunk.

The DjVu standard notes that “the use of the TXTa chunk is discouraged”, the compressed TXTz chunk being preferred (see Self::txtz).

See ZoneBuf for a strongly-typed way to build up data for the zones argument.

pub fn txtz(&mut self, bzz: &[u8]) -> Result<(), Error>

Serialize a TXTz chunk.

See TxtBuf for a strongly-typed way to build up data that can be compressed for the bzz argument.

pub fn djbz(&mut self, jb2: &[u8]) -> Result<(), Error>

Serialize a Djbz chunk.

pub fn sjbz(&mut self, jb2: &[u8]) -> Result<(), Error>

Serialize an Sjbz chunk.

pub fn fg44( &mut self, num_slices: u8, color_space: Iw44ColorSpace, width: u16, height: u16, initial_cdc: u8, iw44: &[u8] ) -> Result<(), Error>

Serialize an FG44 chunk.

initial_cdc must not exceed 127.

pub fn bg44( &mut self, num_slices: u8, color_space: Iw44ColorSpace, width: u16, height: u16, initial_cdc: u8, iw44: &[u8] ) -> Result<SerializeBg44Chunks<'_, 'wr>, Error>

Begin serializing a sequence of BG44 chunks.

initial_cdc must not exceed 127.

pub fn fgbz( &mut self, palette: &[PaletteEntry], indices: Option<(usize, &[u8])> ) -> Result<(), Error>

Serialize an FGbz chunk.

If indices are provided, the first element of the tuple should be the number of indices, and the second element should be the BZZ-compressed indices (pass the output of crate::PaletteIndex::slice_as_bytes to a BZZ compressor).

pub fn incl(&mut self, target_id: &str) -> Result<(), Error>

Serialize an INCL chunk.

pub fn bgjp(&mut self, jpeg: &[u8]) -> Result<(), Error>

Serialize a BGjp chunk.

pub fn fgjp(&mut self, jpeg: &[u8]) -> Result<(), Error>

Serialize an FGjp chunk.